pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "contextsync"
version = "0.1.0"
description = "Change Data Capture for AI-assisted codebases. Keeps context files alive as your code evolves."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "ContextSync", email = "hello@contextsync.dev" },
]
keywords = ["ai", "context", "cdc", "developer-tools", "cursor", "copilot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    # CLI
    "typer[all]>=0.12.0",
    "rich>=13.0",

    # Git
    "gitpython>=3.1.40",

    # AST parsing
    "tree-sitter>=0.23.0",

    # Config & validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",

    # Database
    "sqlalchemy>=2.0.25",
    "aiosqlite>=0.19.0",

    # LLM
    "litellm>=1.30.0",
    "instructor>=1.0.0",
    "httpx>=0.27.0",

    # MCP Server
    "mcp>=1.0.0",

    # Core Utils
    "watchdog>=4.0.0",

    # Templating
    "jinja2>=3.1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1",
    "ruff>=0.3.0",
    "mypy>=1.8",
]

[project.scripts]
contextsync = "contextsync.cli:app"

[project.urls]
Homepage = "https://github.com/contextsync/contextsync"
Documentation = "https://docs.contextsync.dev"
Repository = "https://github.com/contextsync/contextsync"

[tool.hatch.build.targets.wheel]
packages = ["src/contextsync"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
